Move routing-related components to "app" folder
